+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 
 - The `amqp_0_9` output now supports dynamic interpolation functions within the `exchange` field.
 - Field `custom_topic_creation` added to the `kafka` output.
+- New bloblang timestamp method `ts_between`.
 
 ### Fixed